Time: 0 ms (100%), Space: 111.2 MB (72.97%) - LeetHub

+++ 2432-number-of-zero-filled-subarrays/2432-number-of-zero-filled-subarrays.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     long long zeroFilledSubarray(vector<int>& nums) {
-        int res =0 ,count = 0; 
+        long long res =0 ,count = 0; 
         for( auto num : nums){
             if(num == 0) count+=1; 
             else count = 0;
